Fix: bug in selectionlist
If you had a invalid selection for the selectionlist you were trying to
get it would raise an exception and stop the scripts which is not what
we want. If we want all the dagnodes from a selection we don't care
about the rest and the list should be created.